Refactor MotionPhotoReader so that bufferHandler is an instance of a separate BufferHandler class.
